Steam Login fully implemented

+++ app/database/migrations/2013_12_10_171649_create_user_voip_table.php
@@ -15,7 +15,7 @@ class CreateUserVoipTable extends Migration {
 		Schema::create('user_voip', function(Blueprint $table)
 		{
 			$table->increments('id');
-			$table->integer('user_id')
+			$table->bigInteger('user_id')
 				->unsigned();
 			$table->integer('voip_id')
 				->unsigned();
